feat: add support for priority label
