+++ tests/priority.test.ts
@@ -1,31 +1,17 @@
-import { ContextPlugin } from "../src/types/plugin-input";
-import { FormattingEvaluatorModule } from "../src/parser/formatting-evaluator-module";
 import { GitHubIssue } from "../src/github-types";
-import { describe, expect, it, jest } from "@jest/globals";
+import { describe, expect, it } from "@jest/globals";
+import { parsePriorityLabel } from "../src/helpers/github";
 
 describe("FormattingEvaluatorModule", () => {
-  const context = {
-    config: {
-      incentives: {
-        formattingEvaluator: null,
-      },
-    },
-    logger: {
-      error: jest.fn(),
-    },
-  } as unknown as ContextPlugin;
-
-  const module = new FormattingEvaluatorModule(context);
-
   it("should default to priority 1 when no priority label is present", () => {
     const labels: GitHubIssue["labels"] = [];
-    const priority = module["_parsePriorityLabel"](labels);
+    const priority = parsePriorityLabel(labels);
     expect(priority).toBe(1);
   });
 
   it('should return priority 3 when "Priority: 3" label is present', () => {
     const labels = ["Priority: 3"];
-    const priority = module["_parsePriorityLabel"](labels);
+    const priority = parsePriorityLabel(labels);
     expect(priority).toBe(3);
   });
 });
